jax-agents/requirements.txt
# Core dependencies
anthropic>=0.40.0
python-dotenv>=1.0.0
pyyaml>=6.0
pydantic>=2.0.0
rich>=13.0.0
tenacity>=8.0.0
click>=8.0.0

# For running tests
pytest>=7.0.0
jax>=0.4.0
jaxlib>=0.4.0
numpy>=1.24.0

# Development tools (optional)
pytest-cov>=4.0.0
black>=23.0.0
ruff>=0.1.0
mypy>=1.0.0